#1219 split out smvinput.py
